LRAC-14897 extract to an util function
